Refactor and optimize backend

+++ ecommerce-backend/src/main/java/com/abdullah/eCommerce/exceptions/ProductNotFoundException.java
@@ -5,11 +5,10 @@ import lombok.Getter;
 
 @Getter
 public class ProductNotFoundException extends RuntimeException {
-    private final Integer id;
-    public ProductNotFoundException(Integer id) {
-        super("Product not found with id : "+ id);
+    private final Long id;
+    
+    public ProductNotFoundException(Long id) {
+        super("Product not found with id : " + id);
         this.id = id;
     }
-
-
 }
