+++ ecommerce-backend/src/main/java/com/abdullah/eCommerce/repositories/OrderItemRepository.java
@@ -1,7 +1,7 @@
 package com.abdullah.eCommerce.repositories;
 
-import com.abdullah.eCommerce.domain.OrderItem;
+import com.abdullah.eCommerce.entities.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
 }
