+++ ecommerce-backend/src/main/java/com/abdullah/eCommerce/repositories/ProductRepository.java
@@ -1,12 +1,12 @@
 package com.abdullah.eCommerce.repositories;
 
-import com.abdullah.eCommerce.domain.Product;
+import com.abdullah.eCommerce.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-    List<Product> findAllByCategoryId(Integer categoryId);
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    List<Product> findAllByCategoryId(Long categoryId);
 }
